+++ presentation/src/main/java/com/dcs/presentation/core/model/PeopleUiState.kt
@@ -8,7 +8,7 @@ import androidx.paging.PagingData
 import com.dcs.presentation.BuildConfig
 
 data class PersonUiState(
-    val id: Int,
+    val id: Long,
     val name: String,
     val originalName: String,
     val adult: Boolean,
@@ -20,7 +20,7 @@ data class PersonUiState(
 )
 
 data class KnownForUiState(
-    val id: Int,
+    val id: Long,
     val name: String,
     val originalName: String,
     val adult: Boolean,
@@ -80,7 +80,7 @@ class PersonUiStateProvider : PreviewParameterProvider<PagingData<PersonUiState>
             PagingData.from(
                 data = (0 until 10).map {
                     PersonUiState(
-                        id = it,
+                        id = it.toLong(),
                         name = "Person $it",
                         originalName = "Original Person $it",
                         adult = false,
@@ -100,7 +100,7 @@ class PersonUiStateProvider : PreviewParameterProvider<PagingData<PersonUiState>
             PagingData.from(
                 data = (0 until 10).map {
                     PersonUiState(
-                        id = it,
+                        id = it.toLong(),
                         name = "Person $it",
                         originalName = "Original Person $it",
                         adult = false,
@@ -109,7 +109,7 @@ class PersonUiStateProvider : PreviewParameterProvider<PagingData<PersonUiState>
                         profilePath = "https://image.tmdb.org/t/p/w500/4q2NNj4S5c5cUfzjXk5jXgefBvS.jpg",
                         knownFor = (0 until 5).map { idx ->
                             KnownForUiState(
-                                id = idx,
+                                id = idx.toLong(),
                                 originalName = "Original Name $idx",
                                 name = "Name $idx",
                                 adult = false,
